requirements.txt
requests
requests-oauthlib
requests_toolbelt
pyyaml